requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
awscli==1.29.84
pydantic>=2.0.0
typing-extensions>=4.5.0
prometheus-client==0.19.0
psutil==5.9.8
prometheus-fastapi-instrumentator
